Removed Crime from Crystals and added basic logo
Plutonium Core extraction no longer calls the cops.

Also added a credits screen and basic logo to the boot sequence for fun.
Tried to include the names of everyone who's been actively working on
this but if I forgot someone call me a dick and I'll fix it.

Will have a much better boot logo later.